backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
azure-ai-textanalytics==5.3.0
requests==2.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
praw==7.7.1
newsapi-python==0.2.7
schedule==1.2.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2

